feat(telegram): Fix conflicts
